+++ standalone/include/psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_deserialization.h
@@ -141,7 +141,7 @@ std::vector<diagnostic::Message> deserializeMessages(std::istringstream& is);
 }
 
 /**
- * @brief Error indicating a problem during the extraction of the measurement data.
+ * @brief Exception thrown on problems during the extraction of the measurement data.
  */
 class DecodingFailure : public std::runtime_error
 {
@@ -150,7 +150,7 @@ public:
 };
 
 /**
- * @brief Error indicating a problem with the additional field: scan_counter
+ * @brief Exception thrown on problems with the additional field: scan_counter
  *
  * The length specified in the Header of the additional field "scan_counter"
  * must be exactly as defined in NUMBER_OF_BYTES_SCAN_COUNTER for it to be converted.