+++ standalone/include/psen_scan_v2_standalone/data_conversion_layer/raw_processing.h
@@ -34,7 +34,7 @@ namespace data_conversion_layer
 namespace raw_processing
 {
 /**
- * @brief Exception which is thrown if the incoming data from the scanner cannot be processed.
+ * @brief Exception thrown if the incoming data from the scanner cannot be processed.
  */
 class StringStreamFailure : public std::runtime_error
 {