+++ standalone/include/psen_scan_v2_standalone/protocol_layer/scanner_state_machine_def.h
@@ -13,6 +13,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+#include "psen_scan_v2_standalone/data_conversion_layer/start_request_serialization.h"
 namespace psen_scan_v2_standalone
 {
 namespace protocol_layer
@@ -47,6 +48,7 @@ DEFAULT_STATE_IMPL(WaitForStopReply)
 
 DEFAULT_ON_ENTRY_IMPL(Idle)
 
+// \cond Ignore "was not declared or defined" warnings from doxygen
 template <class Event, class FSM>
 void ScannerProtocolDef::Idle::on_exit(Event const&, FSM& fsm)
 {
@@ -98,6 +100,7 @@ void ScannerProtocolDef::Stopped::on_entry(Event const&, FSM& fsm)
 
 DEFAULT_ON_EXIT_IMPL(Stopped)
 
+// \endcond
 //+++++++++++++++++++++++++++++++++ Actions +++++++++++++++++++++++++++++++++++
 
 template <class T>
@@ -111,7 +114,8 @@ inline void ScannerProtocolDef::sendStartRequest(const T& event)
     args_->config_.setHostIp(host_ip.to_ulong());
     PSENSCAN_INFO("StateMachine", "No host ip set! Using local ip: {}", host_ip.to_string());
   }
-  args_->control_client_->write(serialize(data_conversion_layer::start_request::Message(args_->config_)));
+  args_->control_client_->write(
+      data_conversion_layer::start_request::serialize(data_conversion_layer::start_request::Message(args_->config_)));
 }
 
 inline void ScannerProtocolDef::handleStartRequestTimeout(const scanner_events::StartTimeout& event)
@@ -202,9 +206,9 @@ inline void ScannerProtocolDef::sendMessageWithMeasurements(
 inline bool ScannerProtocolDef::framesContainMeasurements(
     const std::vector<data_conversion_layer::monitoring_frame::Message>& frames)
 {
-  if (std::any_of(frames.begin(), frames.end(), [](const auto& frame) { return frame.measurements().empty(); }))
+  if (std::all_of(frames.begin(), frames.end(), [](const auto& frame) { return frame.measurements().empty(); }))
   {
-    PSENSCAN_DEBUG("StateMachine", "No measurement data in this message, skipping laser scan callback.");
+    PSENSCAN_DEBUG("StateMachine", "No measurement data in current monitoring frame(s), skipping laser scan callback.");
     return false;
   }
   return true;