+++ standalone/main.cpp
@@ -42,6 +42,8 @@ int main(int argc, char** argv)
 {
   setLogLevel(CONSOLE_BRIDGE_LOG_INFO);
 
+  // Available configuration options are listed on
+  // http://docs.ros.org/en/melodic/api/psen_scan_v2/html/classpsen__scan__v2__standalone_1_1ScannerConfigurationBuilder.html
   ScannerConfigurationBuilder config_builder;
   config_builder.scannerIp(SCANNER_IP).scanRange(ScanRange{ ANGLE_START, ANGLE_END });