+++ test/include/psen_scan_v2/laserscan_validator.h
@@ -28,8 +28,7 @@
 
 #include "psen_scan_v2_standalone/data_conversion_layer/angle_conversions.h"
 #include "psen_scan_v2_standalone/laserscan.h"
-
-#include "psen_scan_v2_standalone/data_conversion_layer/angle_conversions.h"
+#include "psen_scan_v2_standalone/util/tenth_of_degree.h"
 
 #include "psen_scan_v2/dist.h"
 
@@ -39,8 +38,7 @@ void addScanToBin(const psen_scan_v2_standalone::LaserScan& scan, std::map<int16
 {
   for (size_t i = 0; i < scan.getMeasurements().size(); ++i)
   {
-    auto bin_addr = psen_scan_v2_standalone::data_conversion_layer::radToTenthDegree(scan.getMinScanAngle() +
-                                                                                     scan.getScanResolution() * i);
+    auto bin_addr = (scan.getMinScanAngle() + scan.getScanResolution() * i).value();
 
     if (bin.find(bin_addr) == bin.end())
     {
@@ -56,7 +54,7 @@ void addScanToBin(const sensor_msgs::LaserScan& scan, std::map<int16_t, NormalDi
   for (size_t i = 0; i < scan.ranges.size(); ++i)
   {
     auto bin_addr =
-        psen_scan_v2_standalone::data_conversion_layer::radToTenthDegree(scan->angle_min + scan->angle_increment * i);
+        psen_scan_v2_standalone::data_conversion_layer::radToTenthDegree(scan.angle_min + scan.angle_increment * i);
 
     if (bin.find(bin_addr) == bin.end())
     {